Move @Transactional annotations into separate service
